requirements.txt
torch==2.5.0
torchvision==0.20.0
numpy==1.24.4
pillow==10.4.0
opencv==4.7.0
huggingface-hub==0.29.3
transformers==4.50.2
accelerate==1.5.2
datasets==3.4.0
peft==0.14.0
trl==0.15.2
bitsandbytes==0.45.3
wandb==0.19.8
protobuf==3.19.0
qwen-vl-utils==0.0.10